Color start and end of node conditionally
The node with an input that has the NodePosition Start also runs without something plugged into it. Thus, the handle's green color implies that the input can be safely ignored.
When a node with the NodePosition Middle doesn't have any output attached, the handle is red to communicate that something is probably missing.